doc: Add Apache License 2.0 to the CoP Documentation Project
This commit is to the Apache License 2.0 to the project as the LICENSE file.
The license outlines the terms and conditions for use, reproduction, and
distribution of the project. Including this license ensures that users and
contributors are aware of their rights and responsibilities regarding
the project's code.

This file is to maintain transparency and legal clarity in the use of
the Code of Practice (CoP).

Signed-off-by: Geunsik Lim <leemgs@gmail.com>